Refactor CatalogueContext to add new function getItemsBySellerId
This commit refactors the CatalogueContext by adding a new function
called getItemsBySellerId. This function makes an API call to retrieve
items based on the seller ID. It sets the isLoading state to true while
fetching the items and updates the local state and local storage with
the fetched items. In case of an error, it handles the error
appropriately. This change enables the retrieval of items by seller ID,
enhancing the functionality of the CatalogueContext.