feat: Add authentication to user pages and forms
- Introduce `withAuth` HOC to user-related pages and forms for
authentication purposes
- Ensure that only authenticated users can access and interact with
these pages
- Enhance security and protect user data
- Improves user experience and prevents unauthorized actions
- Resolves issue #123

Note: The 'withAuth' higher-order component (HOC) is responsible for
handling authentication logic and rendering the desired component only
if the user is authenticated. By applying this HOC to the user-related
pages and forms, such as UserAuctions, CatalogueItem, AuctionItemForm,
and UserItems, we ensure that only authenticated users can access and
interact with these pages.